Resolve Circular Dependency (#360)
* Resolve utils --> list curcular dependency

* Remove comment about circular dependency


---------
